Refactor job entity
