+++ pkg/database/database.go
@@ -9,118 +9,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/gosimple/slug"
-	"github.com/lib/pq"
 	"github.com/segmentio/ksuid"
 )
-
-type Job struct {
-	CreatedAt         int64
-	JobTitle          string
-	Company           string
-	SalaryMin         string
-	SalaryMax         string
-	SalaryCurrency    string
-	SalaryPeriod      string
-	SalaryRange       string
-	Location          string
-	Description       string
-	Perks             string
-	InterviewProcess  string
-	HowToApply        string
-	Email             string
-	Expired           bool
-	LastWeekClickouts int
-}
-
-type JobRq struct {
-	JobTitle          string `json:"job_title"`
-	Location          string `json:"job_location"`
-	Company           string `json:"company_name"`
-	CompanyURL        string `json:"company_url"`
-	SalaryMin         string `json:"salary_min"`
-	SalaryMax         string `json:"salary_max"`
-	SalaryCurrency    string `json:"salary_currency"`
-	Description       string `json:"job_description"`
-	HowToApply        string `json:"how_to_apply"`
-	Perks             string `json:"perks"`
-	InterviewProcess  string `json:"interview_process,omitempty"`
-	Email             string `json:"company_email"`
-	StripeToken       string `json:"stripe_token,omitempty"`
-	AdType            int64  `json:"ad_type"`
-	CurrencyCode      string `json:"currency_code"`
-	CompanyIconID     string `json:"company_icon_id,omitempty"`
-	SalaryCurrencyISO string `json:"salary_currency_iso"`
-	VisaSponsorship   bool   `json:"visa_sponsorship,omitempty"`
-}
-
-type JobRqUpsell struct {
-	Token        string `json:"token"`
-	Email        string `json:"email"`
-	StripeToken  string `json:"stripe_token,omitempty"`
-	AdType       int64  `json:"ad_type"`
-	CurrencyCode string `json:"currency_code"`
-}
-
-type JobRqUpdate struct {
-	JobTitle         string `json:"job_title"`
-	Location         string `json:"job_location"`
-	Company          string `json:"company_name"`
-	CompanyURL       string `json:"company_url"`
-	SalaryMin        string `json:"salary_min"`
-	SalaryMax        string `json:"salary_max"`
-	SalaryCurrency   string `json:"salary_currency"`
-	Description      string `json:"job_description"`
-	HowToApply       string `json:"how_to_apply"`
-	Perks            string `json:"perks"`
-	InterviewProcess string `json:"interview_process"`
-	Email            string `json:"company_email"`
-	Token            string `json:"token"`
-	CompanyIconID    string `json:"company_icon_id,omitempty"`
-	SalaryPeriod     string `json:"salary_period"`
-}
-
-type JobPost struct {
-	ID                int
-	CreatedAt         int64
-	TimeAgo           string
-	JobTitle          string
-	Company           string
-	CompanyURL        string
-	SalaryRange       string
-	Location          string
-	JobDescription    string
-	Perks             string
-	InterviewProcess  string
-	HowToApply        string
-	Slug              string
-	SalaryCurrency    string
-	AdType            int64
-	SalaryMin         int64
-	SalaryMax         int64
-	CompanyIconID     string
-	ExternalID        string
-	IsQuickApply      bool
-	ApprovedAt        *time.Time
-	CompanyEmail      string
-	SalaryPeriod      string
-	CompanyURLEnc     string
-	Expired           bool
-	LastWeekClickouts int
-}
-
-type JobPostForEdit struct {
-	ID                                                                        int
-	JobTitle, Company, CompanyEmail, CompanyURL, Location                     string
-	SalaryMin, SalaryMax                                                      int
-	SalaryCurrency, JobDescription, Perks, InterviewProcess, HowToApply, Slug string
-	CreatedAt                                                                 time.Time
-	ApprovedAt                                                                pq.NullTime
-	AdType                                                                    int64
-	CompanyIconID                                                             string
-	ExternalID                                                                string
-	SalaryPeriod                                                              string
-}
 
 type SEOLandingPage struct {
 	URI      string
@@ -144,42 +34,6 @@ type SEOSkill struct {
 //
 // Table Structure:
 //
-// CREATE TABLE IF NOT EXISTS job (
-// 	id        		   SERIAL NOT NULL,
-// 	job_title          VARCHAR(128) NOT NULL,
-// 	company            VARCHAR(128) NOT NULL,
-// 	company_url        VARCHAR(128),
-// 	company_twitter    VARCHAR(128),
-// 	company_email      VARCHAR(128),
-// 	salary_range       VARCHAR(100) NOT NULL,
-// 	location           VARCHAR(200) NOT NULL,
-// 	description        TEXT NOT NULL,
-// 	perks              TEXT,
-// 	interview_process  TEXT,
-// 	how_to_apply       VARCHAR(512),
-// 	created_at         TIMESTAMP NOT NULL,
-// 	approved_at        TIMESTAMP,
-// 	url_id             INTEGER NOT NULL,
-// 	slug               VARCHAR(256),
-//  PRIMARY KEY (id)
-// );
-// CREATE UNIQUE INDEX url_id_idx on job (url_id);
-// CREATE UNIQUE INDEX slug_idx on job (slug);
-// ALTER TABLE job ADD COLUMN salary_min INTEGER NOT NULL DEFAULT 1;
-// ALTER TABLE job ADD COLUMN salary_max INTEGER NOT NULL DEFAULT 1;
-// ALTER TABLE job ADD COLUMN salary_currency VARCHAR(4) NOT NULL DEFAULT '$';
-// ALTER TABLE job ADD COLUMN external_id VARCHAR(28) NOT NULL;
-// ALTER TABLE job ADD COLUMN external_id VARCHAR(28) DROP DEFAULT;
-// ALTER TABLE job ADD COLUMN ad_type INTEGER NOT NULL DEFAULT 0;
-// ALTER TABLE job ALTER COLUMN company_url SET NOT NULL;
-// ALTER TABLE job ADD COLUMN company_icon_image_id VARCHAR(255) DEFAULT NULL;
-// ALTER TABLE job ADD COLUMN salary_period VARCHAR(10) NOT NULL DEFAULT 'year';
-// ALTER TABLE job ADD COLUMN estimated_salary BOOLEAN DEFAULT FALSE;
-// ALTER TABLE job ADD COLUMN expired BOOLEAN DEFAULT FALSE;
-// ALTER TABLE job ADD COLUMN last_week_clickouts INTEGER NOT NULL DEFAULT 0;
-// ALTER TABLE job ADD COLUMN salary_currency_iso CHAR(3) DEFAULT NULL;
-// ALTER TABLE job ADD COLUMN visa_sponsorship BOOLEAN DEFAULT FALSE;
-
 // CREATE TABLE IF NOT EXISTS image (
 // 	id CHAR(27) NOT NULL UNIQUE,
 // 	bytes BYTEA NOT NULL,
@@ -238,14 +92,6 @@ type SEOSkill struct {
 //   updated_at TIMESTAMP NOT NULL,
 //   PRIMARY KEY(base, target)
 // );
-
-// CREATE TABLE IF NOT EXISTS job_event (
-// 	event_type VARCHAR(128) NOT NULL,
-// 	job_id INTEGER NOT NULL REFERENCES job (id),
-// 	created_at TIMESTAMP NOT NULL
-// );
-
-// CREATE INDEX job_idx ON job_event (job_id);
 
 // CREATE TABLE IF NOT EXISTS seo_salary (
 //  id VARCHAR(255) NOT NULL,
@@ -405,15 +251,6 @@ func GetEmailSubscribers(conn *sql.DB) ([]EmailSubscriber, error) {
 	return res, nil
 }
 
-const (
-	jobEventPageView = "page_view"
-	jobEventClickout = "clickout"
-
-	SearchTypeJob     = "job"
-	SearchTypeSalary  = "salary"
-	SearchTypeCompany = "company"
-)
-
 // GetDbConn tries to establish a connection to postgres and return the connection handler
 func GetDbConn(databaseURL string) (*sql.DB, error) {
 	db, err := sql.Open("postgres", databaseURL)
@@ -446,12 +283,6 @@ func DeleteImageByID(conn *sql.DB, id string) error {
 	return err
 }
 
-func TrackJobView(conn *sql.DB, job *JobPost) error {
-	stmt := `INSERT INTO job_event (event_type, job_id, created_at) VALUES ($1, $2, NOW())`
-	_, err := conn.Exec(stmt, jobEventPageView, job.ID)
-	return err
-}
-
 func TrackSearchEvent(conn *sql.DB, ua string, sessionID string, loc string, tag string, results int, typ string) error {
 	hasBot := regexp.MustCompile(`(?i)(googlebot|bingbot|slurp|baiduspider|duckduckbot|yandexbot|sogou|exabot|facebookexternalhit|facebot|ia_archiver|linkedinbot|python-urllib|python-requests|go-http-client|msnbot|ahrefs)`)
 	if hasBot.MatchString(ua) {
@@ -465,102 +296,6 @@ func TrackSearchEvent(conn *sql.DB, ua string, sessionID string, loc string, tag
 	stmt := `INSERT INTO search_event (session_id, location, tag, results, type, created_at) VALUES ($1, NULLIF($2, ''), NULLIF($3, ''), $4, $5, NOW())`
 	_, err := conn.Exec(stmt, sessionID, loc, tag, results, typ)
 	return err
-}
-
-func ApplyToJob(conn *sql.DB, jobID int, cv []byte, email, token string) error {
-	stmt := `INSERT INTO apply_token (token, job_id, created_at, email, cv) VALUES ($1, $2, NOW(), $3, $4)`
-	_, err := conn.Exec(stmt, token, jobID, email, cv)
-	return err
-}
-
-func ConfirmApplyToJob(conn *sql.DB, token string) error {
-	_, err := conn.Exec(
-		`UPDATE apply_token SET confirmed_at = NOW() WHERE token = $1`,
-		token,
-	)
-	return err
-}
-
-type Applicant struct {
-	Cv    []byte
-	Email string
-}
-
-func GetJobByApplyToken(conn *sql.DB, token string) (JobPost, Applicant, error) {
-	res := conn.QueryRow(`SELECT t.cv, t.email, j.id, j.job_title, j.company, company_url, salary_range, location, how_to_apply, slug, j.external_id
-	FROM job j JOIN apply_token t ON t.job_id = j.id AND t.token = $1 WHERE j.approved_at IS NOT NULL AND t.created_at < NOW() + INTERVAL '3 days' AND t.confirmed_at IS NULL`, token)
-	job := JobPost{}
-	applicant := Applicant{}
-	err := res.Scan(&applicant.Cv, &applicant.Email, &job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.HowToApply, &job.Slug, &job.ExternalID)
-	if err != nil {
-		return JobPost{}, applicant, err
-	}
-
-	return job, applicant, nil
-}
-
-func TrackJobClickout(conn *sql.DB, jobID int) error {
-	stmt := `INSERT INTO job_event (event_type, job_id, created_at) VALUES ($1, $2, NOW())`
-	_, err := conn.Exec(stmt, jobEventClickout, jobID)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func GetJobByExternalID(conn *sql.DB, externalID string) (JobPost, error) {
-	res := conn.QueryRow(`SELECT id, job_title, company, company_url, salary_range, location, how_to_apply, slug, external_id, salary_period FROM job WHERE external_id = $1`, externalID)
-	var job JobPost
-	err := res.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.HowToApply, &job.Slug, &job.ExternalID, &job.SalaryPeriod)
-	if err != nil {
-		return job, err
-	}
-
-	return job, nil
-}
-
-type JobAdType int
-
-const (
-	JobAdBasic = iota
-	JobAdSponsoredBackground
-	JobAdSponsoredPinnedFor30Days
-	JobAdSponsoredPinnedFor7Days
-	JobAdWithCompanyLogo
-	JobAdSponsoredPinnedFor60Days
-)
-
-// DemoteJobAdsOlderThan
-func DemoteJobAdsOlderThan(conn *sql.DB, since time.Time, jobAdType JobAdType) (int, error) {
-	res := conn.QueryRow(`WITH rows AS (UPDATE job SET ad_type = $1 WHERE ad_type = $2 AND approved_at <= $3 RETURNING 1) SELECT count(*) as c FROM rows;`, JobAdBasic, jobAdType, since)
-	var affected int
-	err := res.Scan(&affected)
-	if err != nil {
-		return 0, err
-	}
-	return affected, nil
-}
-
-func GetJobsOlderThan(conn *sql.DB, since time.Time, adType JobAdType) ([]JobPost, error) {
-	var jobs []JobPost
-	rows, err := conn.Query(`SELECT id, job_title, company, company_url, company_email, salary_range, location, how_to_apply, slug, external_id, approved_at FROM job j WHERE approved_at <= $1 AND ad_type = $2`, since, adType)
-	if err == sql.ErrNoRows {
-		return jobs, nil
-	}
-	for rows.Next() {
-		var job JobPost
-		var approvedAt sql.NullTime
-		err := rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.CompanyEmail, &job.SalaryRange, &job.Location, &job.HowToApply, &job.Slug, &job.ExternalID, &approvedAt)
-		if err != nil {
-			return jobs, err
-		}
-		if approvedAt.Valid {
-			job.ApprovedAt = &approvedAt.Time
-		}
-		jobs = append(jobs, job)
-	}
-
-	return jobs, nil
 }
 
 func LocationsByPrefix(conn *sql.DB, prefix string) ([]Location, error) {
@@ -609,11 +344,6 @@ func SkillsByPrefix(conn *sql.DB, prefix string) ([]SEOSkill, error) {
 	}
 
 	return skills, nil
-}
-
-func UpdateJobAdType(conn *sql.DB, adType int, jobID int) error {
-	_, err := conn.Exec(`UPDATE job SET ad_type = $1, approved_at = NOW() WHERE id = $2`, adType, jobID)
-	return err
 }
 
 type SalaryDataPoint struct {
@@ -834,131 +564,6 @@ func GetSEOskills(conn *sql.DB) ([]SEOSkill, error) {
 	return skills, nil
 }
 
-func SaveDraft(db *sql.DB, job *JobRq) (int, error) {
-	externalID, err := ksuid.NewRandom()
-	if err != nil {
-		return 0, err
-	}
-	sqlStatement := `
-			INSERT INTO job (job_title, company, company_url, salary_range, salary_min, salary_max, salary_currency, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, company_email, ad_type, external_id, salary_period, salary_currency_iso, visa_sponsorship)
-			VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, 'year', $19, $20) RETURNING id`
-	if job.CompanyIconID != "" {
-		sqlStatement = `
-			INSERT INTO job (job_title, company, company_url, salary_range, salary_min, salary_max, salary_currency, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, company_email, ad_type, company_icon_image_id, external_id, salary_period, salary_currency_iso, visa_sponsorship)
-			VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, 'year', $20, $21) RETURNING id`
-	}
-	slugTitle := slug.Make(fmt.Sprintf("%s %s %d", job.JobTitle, job.Company, time.Now().UTC().Unix()))
-	createdAt := time.Now().UTC().Unix()
-	salaryMinInt, err := strconv.Atoi(strings.TrimSpace(job.SalaryMin))
-	if err != nil {
-		return 0, err
-	}
-	salaryMaxInt, err := strconv.Atoi(strings.TrimSpace(job.SalaryMax))
-	if err != nil {
-		return 0, err
-	}
-	salaryRange := SalaryToSalaryRangeString(salaryMinInt, salaryMaxInt, job.SalaryCurrency)
-	var lastInsertID int
-	var res *sql.Row
-	if job.CompanyIconID != "" {
-		res = db.QueryRow(sqlStatement, job.JobTitle, job.Company, job.CompanyURL, salaryRange, job.SalaryMin, job.SalaryMax, job.SalaryCurrency, job.Location, job.Description, job.Perks, job.InterviewProcess, job.HowToApply, time.Unix(createdAt, 0), createdAt, slugTitle, job.Email, job.AdType, job.CompanyIconID, externalID, job.SalaryCurrencyISO, job.VisaSponsorship)
-	} else {
-		res = db.QueryRow(sqlStatement, job.JobTitle, job.Company, job.CompanyURL, salaryRange, job.SalaryMin, job.SalaryMax, job.SalaryCurrency, job.Location, job.Description, job.Perks, job.InterviewProcess, job.HowToApply, time.Unix(createdAt, 0), createdAt, slugTitle, job.Email, job.AdType, externalID, job.SalaryCurrencyISO, job.VisaSponsorship)
-	}
-	res.Scan(&lastInsertID)
-	if err != nil {
-		return 0, err
-	}
-	return int(lastInsertID), err
-}
-
-func UpdateJob(conn *sql.DB, job *JobRqUpdate, jobID int) error {
-	salaryMinInt, err := strconv.Atoi(strings.TrimSpace(job.SalaryMin))
-	if err != nil {
-		return err
-	}
-	salaryMaxInt, err := strconv.Atoi(strings.TrimSpace(job.SalaryMax))
-	if err != nil {
-		return err
-	}
-	salaryRange := SalaryToSalaryRangeString(salaryMinInt, salaryMaxInt, job.SalaryCurrency)
-	_, err = conn.Exec(
-		`UPDATE job SET job_title = $1, company = $2, company_url = $3, salary_min = $4, salary_max = $5, salary_currency = $6, salary_range = $7, location = $8, description = $9, perks = $10, interview_process = $11, how_to_apply = $12, company_icon_image_id = $13 WHERE id = $14`,
-		job.JobTitle,
-		job.Company,
-		job.CompanyURL,
-		job.SalaryMin,
-		job.SalaryMax,
-		job.SalaryCurrency,
-		salaryRange,
-		job.Location,
-		job.Description,
-		job.Perks,
-		job.InterviewProcess,
-		job.HowToApply,
-		job.CompanyIconID,
-		jobID,
-	)
-	if err != nil {
-		return err
-	}
-	return err
-}
-
-func ApproveJob(conn *sql.DB, jobID int) error {
-	_, err := conn.Exec(
-		`UPDATE job SET approved_at = NOW() WHERE id = $1`,
-		jobID,
-	)
-	if err != nil {
-		return err
-	}
-	return err
-}
-
-func DisapproveJob(conn *sql.DB, jobID int) error {
-	_, err := conn.Exec(
-		`UPDATE job SET approved_at = NULL WHERE id = $1`,
-		jobID,
-	)
-	if err != nil {
-		return err
-	}
-	return err
-}
-
-func SalaryToSalaryRangeString(salaryMin, salaryMax int, currency string) string {
-	salaryMinStr := fmt.Sprintf("%d", salaryMin)
-	salaryMaxStr := fmt.Sprintf("%d", salaryMax)
-	if currency != "₹" {
-		if salaryMin > 1000 {
-			salaryMinStr = fmt.Sprintf("%dk", salaryMin/1000)
-		}
-		if salaryMax > 1000 {
-			salaryMaxStr = fmt.Sprintf("%dk", salaryMax/1000)
-		}
-	} else {
-		if salaryMin > 100000 {
-			salaryMinStr = fmt.Sprintf("%dL", salaryMin/100000)
-		}
-		if salaryMax > 100000 {
-			salaryMaxStr = fmt.Sprintf("%dL", salaryMax/100000)
-		}
-	}
-
-	return fmt.Sprintf("%s%s - %s%s", currency, salaryMinStr, currency, salaryMaxStr)
-}
-
-func GetViewCountForJob(conn *sql.DB, jobID int) (int, error) {
-	var count int
-	row := conn.QueryRow(`select count(*) as c from job_event where job_event.event_type = 'page_view' and job_event.job_id = $1`, jobID)
-	err := row.Scan(&count)
-	if err != nil {
-		return 0, err
-	}
-	return count, err
-}
-
 type PurchaseEvent struct {
 	StripeSessionID string
 	CreatedAt       time.Time
@@ -1016,699 +621,6 @@ func GetPurchaseEventBySessionID(conn *sql.DB, sessionID string) (PurchaseEvent,
 	}
 
 	return p, nil
-}
-
-func GetJobByStripeSessionID(conn *sql.DB, sessionID string) (JobPost, error) {
-	res := conn.QueryRow(`SELECT j.id, j.job_title, j.company, j.company_url, j.salary_range, j.location, j.how_to_apply, j.slug, j.external_id, j.approved_at FROM purchase_event p LEFT JOIN job j ON p.job_id = j.id WHERE p.stripe_session_id = $1`, sessionID)
-	var job JobPost
-	var approvedAt sql.NullTime
-	err := res.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.HowToApply, &job.Slug, &job.ExternalID, &approvedAt)
-	if err != nil {
-		return job, err
-	}
-	if approvedAt.Valid {
-		job.ApprovedAt = &approvedAt.Time
-	}
-
-	return job, nil
-}
-
-type JobStat struct {
-	Date      string `json:"date"`
-	Clickouts int    `json:"clickouts"`
-	PageViews int    `json:"pageviews"`
-}
-
-func GetStatsForJob(conn *sql.DB, jobID int) ([]JobStat, error) {
-	var stats []JobStat
-	rows, err := conn.Query(`SELECT COUNT(*) FILTER (WHERE event_type = 'clickout') AS clickout, COUNT(*) FILTER (WHERE event_type = 'page_view') AS pageview, TO_CHAR(DATE_TRUNC('day', created_at), 'YYYY-MM-DD') FROM job_event WHERE job_id = $1 GROUP BY DATE_TRUNC('day', created_at) ORDER BY DATE_TRUNC('day', created_at) ASC`, jobID)
-	if err == sql.ErrNoRows {
-		return stats, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-	for rows.Next() {
-		var s JobStat
-		if err := rows.Scan(&s.Clickouts, &s.PageViews, &s.Date); err != nil {
-			return stats, err
-		}
-		stats = append(stats, s)
-	}
-
-	return stats, nil
-}
-
-func GetClickoutCountForJob(conn *sql.DB, jobID int) (int, error) {
-	var count int
-	row := conn.QueryRow(`select count(*) as c from job_event where job_event.event_type = 'clickout' and job_event.job_id = $1`, jobID)
-	err := row.Scan(&count)
-	if err != nil {
-		return 0, err
-	}
-	return count, err
-}
-
-func JobPostByCreatedAt(conn *sql.DB) ([]*JobPost, error) {
-	jobs := []*JobPost{}
-	var rows *sql.Rows
-	rows, err := conn.Query(
-		`SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-		FROM job
-		WHERE approved_at IS NOT NULL
-		ORDER BY created_at DESC`)
-	if err != nil {
-		return jobs, err
-	}
-	for rows.Next() {
-		job := &JobPost{}
-		var createdAt time.Time
-		var perks, interview, companyIcon sql.NullString
-		err = rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod, &job.Expired, &job.LastWeekClickouts)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if perks.Valid {
-			job.Perks = perks.String
-		}
-		if interview.Valid {
-			job.InterviewProcess = interview.String
-		}
-		job.TimeAgo = createdAt.UTC().Format("January 2006")
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	err = rows.Err()
-	if err != nil {
-		return jobs, err
-	}
-	return jobs, nil
-}
-
-func TopNJobsByCurrencyAndLocation(conn *sql.DB, currency, location string, max int) ([]*JobPost, error) {
-	jobs := []*JobPost{}
-	var rows *sql.Rows
-	rows, err := conn.Query(
-		`SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period
-		FROM job
-		WHERE salary_currency = $1
-		AND location ILIKE '%' || $2 || '%'
-		AND approved_at IS NOT NULL
-		ORDER BY created_at DESC LIMIT $3`, currency, location, max)
-	if err != nil {
-		return jobs, err
-	}
-	for rows.Next() {
-		job := &JobPost{}
-		var createdAt time.Time
-		var perks, interview, companyIcon sql.NullString
-		err = rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if perks.Valid {
-			job.Perks = perks.String
-		}
-		if interview.Valid {
-			job.InterviewProcess = interview.String
-		}
-		job.TimeAgo = createdAt.UTC().Format("January 2006")
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	err = rows.Err()
-	if err != nil {
-		return jobs, err
-	}
-	return jobs, nil
-}
-
-func JobPostBySlug(conn *sql.DB, slug string) (*JobPost, error) {
-	job := &JobPost{}
-	row := conn.QueryRow(
-		`SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-		FROM job
-		WHERE approved_at IS NOT NULL
-		AND slug = $1`, slug)
-	var createdAt time.Time
-	var perks, interview, companyIcon sql.NullString
-	err := row.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod, &job.Expired, &job.LastWeekClickouts)
-	if companyIcon.Valid {
-		job.CompanyIconID = companyIcon.String
-	}
-	if err != nil {
-		return job, err
-	}
-	if perks.Valid {
-		job.Perks = perks.String
-	}
-	if interview.Valid {
-		job.InterviewProcess = interview.String
-	}
-	job.TimeAgo = createdAt.UTC().Format("January 2006")
-	return job, nil
-}
-
-func JobPostBySlugAdmin(conn *sql.DB, slug string) (*JobPost, error) {
-	job := &JobPost{}
-	row := conn.QueryRow(
-		`SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period
-		FROM job
-		WHERE slug = $1`, slug)
-	var createdAt time.Time
-	var perks, interview, companyIcon sql.NullString
-	err := row.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod)
-	if companyIcon.Valid {
-		job.CompanyIconID = companyIcon.String
-	}
-	if err != nil {
-		return job, err
-	}
-	if perks.Valid {
-		job.Perks = perks.String
-	}
-	if interview.Valid {
-		job.InterviewProcess = interview.String
-	}
-	job.TimeAgo = createdAt.UTC().Format("January 2006")
-	return job, nil
-}
-
-func JobPostByIDForEdit(conn *sql.DB, jobID int) (*JobPostForEdit, error) {
-	job := &JobPostForEdit{}
-	row := conn.QueryRow(
-		`SELECT job_title, company, company_email, company_url, salary_min, salary_max, salary_currency, location, description, perks, interview_process, how_to_apply, created_at, slug, approved_at, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period
-		FROM job
-		WHERE id = $1`, jobID)
-	var perks, interview, companyURL, companyIconID sql.NullString
-	err := row.Scan(&job.JobTitle, &job.Company, &job.CompanyEmail, &companyURL, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &job.CreatedAt, &job.Slug, &job.ApprovedAt, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIconID, &job.ExternalID, &job.SalaryPeriod)
-	if err != nil {
-		return job, err
-	}
-	if companyIconID.Valid {
-		job.CompanyIconID = companyIconID.String
-	}
-	if perks.Valid {
-		job.Perks = perks.String
-	}
-	if interview.Valid {
-		job.InterviewProcess = interview.String
-	}
-	if companyURL.Valid {
-		job.CompanyURL = companyURL.String
-	} else {
-		job.CompanyURL = ""
-	}
-	return job, nil
-}
-
-func JobPostByExternalIDForEdit(conn *sql.DB, externalID string) (*JobPostForEdit, error) {
-	job := &JobPostForEdit{}
-	row := conn.QueryRow(
-		`SELECT id, job_title, company, company_email, company_url, salary_min, salary_max, salary_currency, location, description, perks, interview_process, how_to_apply, created_at, slug, approved_at, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period
-		FROM job
-		WHERE external_id = $1`, externalID)
-	var perks, interview, companyURL, companyIconID sql.NullString
-	err := row.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyEmail, &companyURL, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &job.CreatedAt, &job.Slug, &job.ApprovedAt, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIconID, &job.ExternalID, &job.SalaryPeriod)
-	if err != nil {
-		return job, err
-	}
-	if companyIconID.Valid {
-		job.CompanyIconID = companyIconID.String
-	}
-	if perks.Valid {
-		job.Perks = perks.String
-	}
-	if interview.Valid {
-		job.InterviewProcess = interview.String
-	}
-	if companyURL.Valid {
-		job.CompanyURL = companyURL.String
-	} else {
-		job.CompanyURL = ""
-	}
-	return job, nil
-}
-
-func JobPostByURLID(conn *sql.DB, URLID int64) (*JobPost, error) {
-	job := &JobPost{}
-	row := conn.QueryRow(
-		`SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-		FROM job
-		WHERE approved_at IS NOT NULL
-		AND url_id = $1`, URLID)
-	var createdAt time.Time
-	var perks, interview, companyIcon sql.NullString
-	err := row.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod, &job.Expired, &job.LastWeekClickouts)
-	if err != nil {
-		return job, err
-	}
-	if companyIcon.Valid {
-		job.CompanyIconID = companyIcon.String
-	}
-	if perks.Valid {
-		job.Perks = perks.String
-	}
-	if interview.Valid {
-		job.InterviewProcess = interview.String
-	}
-	job.TimeAgo = createdAt.UTC().Format("January 2006")
-	return job, nil
-}
-
-func DeleteJobCascade(conn *sql.DB, jobID int) error {
-	if _, err := conn.Exec(
-		`DELETE FROM image WHERE id IN (SELECT company_icon_image_id FROM job WHERE id = $1)`,
-		jobID,
-	); err != nil {
-		return err
-	}
-	if _, err := conn.Exec(
-		`DELETE FROM edit_token WHERE job_id = $1`,
-		jobID,
-	); err != nil {
-		return err
-	}
-	if _, err := conn.Exec(
-		`DELETE FROM apply_token WHERE job_id = $1`,
-		jobID,
-	); err != nil {
-		return err
-	}
-	if _, err := conn.Exec(
-		`DELETE FROM job_event WHERE job_id = $1`,
-		jobID,
-	); err != nil {
-		return err
-	}
-	if _, err := conn.Exec(
-		`DELETE FROM purchase_event WHERE job_id = $1`,
-		jobID,
-	); err != nil {
-		return err
-	}
-	if _, err := conn.Exec(
-		`DELETE FROM job WHERE id = $1`,
-		jobID,
-	); err != nil {
-		return err
-	}
-	return nil
-}
-
-func GetPendingJobs(conn *sql.DB) ([]*JobPost, error) {
-	jobs := []*JobPost{}
-	var rows *sql.Rows
-	rows, err := conn.Query(`
-	SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period
-		FROM job WHERE approved_at IS NULL`)
-	if err == sql.ErrNoRows {
-		return jobs, nil
-	}
-	if err != nil {
-		return jobs, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		job := &JobPost{}
-		var createdAt time.Time
-		var perks, interview, companyIcon sql.NullString
-		err = rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if perks.Valid {
-			job.Perks = perks.String
-		}
-		if interview.Valid {
-			job.InterviewProcess = interview.String
-		}
-		job.TimeAgo = createdAt.UTC().Format("January 2006")
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	err = rows.Err()
-	if err != nil {
-		return jobs, err
-	}
-	return jobs, nil
-}
-
-// GetCompanyJobs returns jobs for a given company
-func GetCompanyJobs(conn *sql.DB, companyName string, limit int) ([]*JobPost, error) {
-	jobs := []*JobPost{}
-	var rows *sql.Rows
-	rows, err := conn.Query(`
-	SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, last_week_clickouts
-		FROM job WHERE approved_at IS NOT NULL AND expired IS FALSE AND company = $1 ORDER BY ad_type DESC, approved_at DESC LIMIT $2`, companyName, limit)
-	if err != nil {
-		return jobs, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		job := &JobPost{}
-		var createdAt time.Time
-		var perks, interview, companyIcon sql.NullString
-		err = rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod, &job.LastWeekClickouts)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if perks.Valid {
-			job.Perks = perks.String
-		}
-		if interview.Valid {
-			job.InterviewProcess = interview.String
-		}
-		job.TimeAgo = createdAt.UTC().Format("January 2006")
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	err = rows.Err()
-	if err != nil {
-		return jobs, err
-	}
-	return jobs, nil
-}
-
-// GetRelevantJobs returns pinned and most recent jobs for now
-func GetRelevantJobs(conn *sql.DB, location string, jobID int, limit int) ([]*JobPost, error) {
-	jobs := []*JobPost{}
-	var rows *sql.Rows
-	rows, err := conn.Query(`
-	SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, last_week_clickouts
-		FROM job WHERE approved_at IS NOT NULL AND id != $1 AND expired IS FALSE ORDER BY ad_type DESC, approved_at DESC, word_similarity($2, location) LIMIT $3`, jobID, location, limit)
-	if err != nil {
-		return jobs, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		job := &JobPost{}
-		var createdAt time.Time
-		var perks, interview, companyIcon sql.NullString
-		err = rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod, &job.LastWeekClickouts)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if perks.Valid {
-			job.Perks = perks.String
-		}
-		if interview.Valid {
-			job.InterviewProcess = interview.String
-		}
-		job.TimeAgo = createdAt.UTC().Format("January 2006")
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	err = rows.Err()
-	if err != nil {
-		return jobs, err
-	}
-	return jobs, nil
-}
-
-func GetPinnedJobs(conn *sql.DB) ([]*JobPost, error) {
-	jobs := []*JobPost{}
-	var rows *sql.Rows
-	rows, err := conn.Query(`
-	SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, last_week_clickouts
-		FROM job WHERE approved_at IS NOT NULL AND ad_type IN (2, 3, 5) ORDER BY approved_at DESC`)
-	if err != nil {
-		return jobs, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		job := &JobPost{}
-		var createdAt time.Time
-		var perks, interview, companyIcon sql.NullString
-		err = rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod, &job.LastWeekClickouts)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if perks.Valid {
-			job.Perks = perks.String
-		}
-		if interview.Valid {
-			job.InterviewProcess = interview.String
-		}
-		job.TimeAgo = createdAt.UTC().Format("January 2006")
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	err = rows.Err()
-	if err != nil {
-		return jobs, err
-	}
-	return jobs, nil
-}
-
-func JobsByQuery(conn *sql.DB, location, tag string, pageId, salary int, currency string, jobsPerPage int, includePinnedJobs bool) ([]*JobPost, int, error) {
-	jobs := []*JobPost{}
-	var rows *sql.Rows
-	offset := pageId*jobsPerPage - jobsPerPage
-	// replace `|` with white space
-	// remove double white spaces
-	// join with `|` for ps query
-	tag = strings.Join(strings.Fields(strings.ReplaceAll(tag, "|", " ")), "|")
-	rows, err := getQueryForArgs(conn, location, tag, salary, currency, offset, jobsPerPage, includePinnedJobs)
-	if err != nil {
-		return jobs, 0, err
-	}
-	defer rows.Close()
-	var fullRowsCount int
-	for rows.Next() {
-		job := &JobPost{}
-		var createdAt time.Time
-		var perks, interview, companyIcon sql.NullString
-		err = rows.Scan(&fullRowsCount, &job.ID, &job.JobTitle, &job.Company, &job.CompanyURL, &job.SalaryRange, &job.Location, &job.JobDescription, &perks, &interview, &job.HowToApply, &createdAt, &job.CreatedAt, &job.Slug, &job.AdType, &job.SalaryMin, &job.SalaryMax, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod, &job.Expired, &job.LastWeekClickouts)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if perks.Valid {
-			job.Perks = perks.String
-		}
-		if interview.Valid {
-			job.InterviewProcess = interview.String
-		}
-		job.TimeAgo = createdAt.UTC().Format("January 2006")
-		if err != nil {
-			return jobs, fullRowsCount, err
-		}
-		jobs = append(jobs, job)
-	}
-	err = rows.Err()
-	if err != nil {
-		return jobs, fullRowsCount, err
-	}
-	return jobs, fullRowsCount, nil
-}
-
-func TokenByJobID(conn *sql.DB, jobID int) (string, error) {
-	tokenRow := conn.QueryRow(
-		`SELECT token
-		FROM edit_token
-		WHERE job_id = $1`, jobID)
-	var token string
-	err := tokenRow.Scan(&token)
-	return token, err
-}
-
-func JobPostIDByToken(conn *sql.DB, token string) (int, error) {
-	row := conn.QueryRow(
-		`SELECT job_id
-		FROM edit_token
-		WHERE token = $1`, token)
-	var jobID int
-	err := row.Scan(&jobID)
-	if err != nil {
-		return 0, err
-	}
-	return jobID, nil
-}
-
-func LastJobPosted(conn *sql.DB) (time.Time, error) {
-	row := conn.QueryRow(`SELECT created_at FROM job WHERE approved_at IS NOT NULL ORDER BY created_at DESC LIMIT 1`)
-	var last time.Time
-	if err := row.Scan(&last); err != nil {
-		return last, err
-	}
-
-	return last, nil
-}
-
-func SaveTokenForJob(conn *sql.DB, token string, jobID int) error {
-	_, err := conn.Exec(`INSERT INTO edit_token (token, job_id, created_at) VALUES ($1, $2, $3)`, token, jobID, time.Now().UTC())
-	if err != nil {
-		return err
-	}
-	return err
-}
-
-func getQueryForArgs(conn *sql.DB, location, tag string, salary int, currency string, offset, max int, includePinnedJobs bool) (*sql.Rows, error) {
-	adTypeFilter := "AND ad_type NOT IN (2, 3, 5)"
-	if includePinnedJobs {
-		adTypeFilter = "AND 1=1"
-	}
-	if tag == "" && location == "" && salary == 0 {
-		return conn.Query(`
-		SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-		FROM job
-		WHERE approved_at IS NOT NULL `+adTypeFilter+` ORDER BY created_at DESC LIMIT $2 OFFSET $1`, offset, max)
-	}
-	if tag == "" && location != "" && salary == 0 {
-		return conn.Query(`
-		SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts 
-		FROM job
-		WHERE approved_at IS NOT NULL `+adTypeFilter+` AND location ILIKE '%' || $1 || '%'
-		ORDER BY created_at DESC LIMIT $3 OFFSET $2`, location, offset, max)
-	}
-	if tag != "" && location == "" && salary == 0 {
-		return conn.Query(`
-	SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-	FROM
-	(
-		SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts, to_tsvector(job_title) || to_tsvector(company) || to_tsvector(description) AS doc
-		FROM job WHERE approved_at IS NOT NULL `+adTypeFilter+`
-	) AS job_
-	WHERE job_.doc @@ to_tsquery($1)
-	ORDER BY ts_rank(job_.doc, to_tsquery($1)) DESC, created_at DESC LIMIT $3 OFFSET $2`, tag, offset, max)
-	}
-	if tag != "" && location != "" && salary == 0 {
-		return conn.Query(`
-	SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-	FROM
-	(
-		SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts, to_tsvector(job_title) || to_tsvector(company) || to_tsvector(description) AS doc
-		FROM job WHERE approved_at IS NOT NULL `+adTypeFilter+`
-	) AS job_
-	WHERE job_.doc @@ to_tsquery($1)
-	AND location ILIKE '%' || $2 || '%'
-	ORDER BY ts_rank(job_.doc, to_tsquery($1)) DESC, created_at DESC LIMIT $4 OFFSET $3`, tag, location, offset, max)
-	}
-	if tag == "" && location == "" && salary != 0 {
-		return conn.Query(`
-		SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-		FROM job FULL JOIN fx_rate ON fx_rate.base = job.salary_currency_iso AND fx_rate.target = $4
-		WHERE approved_at IS NOT NULL `+adTypeFilter+` AND (COALESCE(fx_rate.value, 1)*job.salary_max) >= $3 ORDER BY created_at DESC LIMIT $2 OFFSET $1`, offset, max, salary, currency)
-	}
-	if tag == "" && location != "" && salary != 0 {
-		return conn.Query(`
-		SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts 
-		FROM job FULL JOIN fx_rate ON fx_rate.base = job.salary_currency_iso AND fx_rate.target = $5
-		WHERE approved_at IS NOT NULL `+adTypeFilter+` AND location ILIKE '%' || $1 || '%' AND (COALESCE(fx_rate.value, 1)*job.salary_max) >= $4
-		ORDER BY created_at DESC LIMIT $3 OFFSET $2`, location, offset, max, salary, currency)
-	}
-	if tag != "" && location == "" && salary != 0 {
-		return conn.Query(`
-	SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-	FROM
-	(
-		SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts, to_tsvector(job_title) || to_tsvector(company) || to_tsvector(description) AS doc
-		FROM job FULL JOIN fx_rate ON fx_rate.base = job.salary_currency_iso AND fx_rate.target = $5 WHERE approved_at IS NOT NULL `+adTypeFilter+` AND (COALESCE(fx_rate.value, 1)*job.salary_max) >= $4
-	) AS job_
-	WHERE job_.doc @@ to_tsquery($1)
-	ORDER BY ts_rank(job_.doc, to_tsquery($1)) DESC, created_at DESC LIMIT $3 OFFSET $2`, tag, offset, max, salary, currency)
-	}
-
-	return conn.Query(`
-	SELECT count(*) OVER() AS full_count, id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts
-	FROM
-	(
-		SELECT id, job_title, company, company_url, salary_range, location, description, perks, interview_process, how_to_apply, created_at, url_id, slug, ad_type, salary_min, salary_max, salary_currency, company_icon_image_id, external_id, salary_period, expired, last_week_clickouts, to_tsvector(job_title) || to_tsvector(company) || to_tsvector(description) AS doc
-		FROM job FULL JOIN fx_rate ON fx_rate.base = job.salary_currency_iso AND fx_rate.target = $6 WHERE approved_at IS NOT NULL `+adTypeFilter+` AND (COALESCE(fx_rate.value, 1)*job.salary_max) >= $5
-	) AS job_
-	WHERE job_.doc @@ to_tsquery($1)
-	AND location ILIKE '%' || $2 || '%'
-	ORDER BY ts_rank(job_.doc, to_tsquery($1)) DESC, created_at DESC LIMIT $4 OFFSET $3`, tag, location, offset, max, salary, currency)
-}
-
-func GetValue(conn *sql.DB, key string) (string, error) {
-	res := conn.QueryRow(`SELECT value FROM meta WHERE key = $1`, key)
-	var val string
-	err := res.Scan(&val)
-	if err != nil {
-		return "", err
-	}
-	return val, nil
-}
-
-func SetValue(conn *sql.DB, key, val string) error {
-	_, err := conn.Exec(`UPDATE meta SET value = $1 WHERE key = $2`, val, key)
-	return err
-}
-
-func GetLastNJobs(conn *sql.DB, max int, loc string) ([]*JobPost, error) {
-	var jobs []*JobPost
-	var rows *sql.Rows
-	var err error
-	if strings.TrimSpace(loc) == "" {
-		rows, err = conn.Query(`SELECT id, job_title, description, company, salary_range, location, slug, salary_currency, company_icon_image_id, external_id, approved_at, salary_period FROM job WHERE approved_at IS NOT NULL ORDER BY approved_at DESC LIMIT $1`, max)
-	} else {
-		rows, err = conn.Query(`SELECT
-	id, job_title, description, company, salary_range, location, slug, salary_currency, company_icon_image_id, external_id, approved_at, salary_period
-	FROM
-	job
-	WHERE
-	approved_at IS NOT NULL
-	AND location ILIKE '%' || $1 || '%'
-	ORDER BY approved_at DESC LIMIT $2`, loc, max)
-	}
-	if err != nil {
-		return jobs, err
-	}
-	for rows.Next() {
-		job := &JobPost{}
-		var companyIcon sql.NullString
-		err := rows.Scan(&job.ID, &job.JobTitle, &job.JobDescription, &job.Company, &job.SalaryRange, &job.Location, &job.Slug, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.ApprovedAt, &job.SalaryPeriod)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	return jobs, nil
-}
-
-func GetLastNJobsFromID(conn *sql.DB, max, jobID int) ([]*JobPost, error) {
-	var jobs []*JobPost
-	var rows *sql.Rows
-	rows, err := conn.Query(`SELECT id, job_title, company, salary_range, location, slug, salary_currency, company_icon_image_id, external_id, salary_period FROM job WHERE id > $1 AND approved_at IS NOT NULL LIMIT $2`, jobID, max)
-	if err != nil {
-		return jobs, err
-	}
-	for rows.Next() {
-		job := &JobPost{}
-		var companyIcon sql.NullString
-		err := rows.Scan(&job.ID, &job.JobTitle, &job.Company, &job.SalaryRange, &job.Location, &job.Slug, &job.SalaryCurrency, &companyIcon, &job.ExternalID, &job.SalaryPeriod)
-		if companyIcon.Valid {
-			job.CompanyIconID = companyIcon.String
-		}
-		if err != nil {
-			return jobs, err
-		}
-		jobs = append(jobs, job)
-	}
-	return jobs, nil
-}
-
-func CleanupExpiredApplyTokens(conn *sql.DB) error {
-	_, err := conn.Exec(
-		`DELETE FROM apply_token WHERE created_at < NOW() - INTERVAL '3 days' OR confirmed_at IS NOT NULL`,
-	)
-	return err
 }
 
 type Media struct {
@@ -1838,20 +750,6 @@ func GetSitemapEntryCount(conn *sql.DB) (int, error) {
 	return count, nil
 }
 
-func NewJobsLastWeekOrMonth(conn *sql.DB) (int, int, error) {
-	var week, month int
-	row := conn.QueryRow(`select lastweek.c as week, lastmonth.c as month 
-from 
-(select count(*) as c, 1 as id from job  where approved_at >= (now() - '7 days'::interval)::date) as lastweek
-left join 
-(select count(*) as c, 1 as id from job  where approved_at >= (now() - '30 days'::interval)::date) as lastmonth on lastmonth.id = lastweek.id 
-`)
-	if err := row.Scan(&week, &month); err != nil {
-		return week, month, err
-	}
-	return week, month, nil
-}
-
 func GetSitemapLastMod(conn *sql.DB) (time.Time, error) {
 	var lastMod time.Time
 	row := conn.QueryRow(`SELECT lastmod FROM sitemap ORDER BY lastmod DESC LIMIT 1`)
@@ -1946,35 +844,8 @@ func SwapSitemapTable(conn *sql.DB) error {
 	return err
 }
 
-func MarkJobAsExpired(conn *sql.DB, jobID int) error {
-	_, err := conn.Exec(`UPDATE job SET expired = true WHERE id = $1`, jobID)
-	return err
-}
-
 func UpdateLastWeekClickouts(conn *sql.DB) error {
 	_, err := conn.Exec(`WITH cte AS (SELECT job_id, count(*) AS clickouts FROM job_event WHERE event_type = 'clickout' AND created_at > CURRENT_DATE - 7 GROUP BY job_id)
 UPDATE job SET last_week_clickouts = cte.clickouts FROM cte WHERE cte.job_id = id`)
 	return err
-}
-
-type JobApplyURL struct {
-	ID  int
-	URL string
-}
-
-func GetJobApplyURLs(conn *sql.DB) ([]JobApplyURL, error) {
-	jobURLs := make([]JobApplyURL, 0)
-	var rows *sql.Rows
-	rows, err := conn.Query(`SELECT id, how_to_apply FROM job WHERE approved_at IS NOT NULL AND expired = false`)
-	if err != nil {
-		return jobURLs, err
-	}
-	for rows.Next() {
-		jobURL := JobApplyURL{}
-		if err := rows.Scan(&jobURL.ID, &jobURL.URL); err != nil {
-			return jobURLs, err
-		}
-		jobURLs = append(jobURLs, jobURL)
-	}
-	return jobURLs, nil
 }
